Moved IonCliCommand trait to `commands` module.
