+++ src/bin/ion/commands/beta/from/mod.rs
@@ -1,4 +1,4 @@
-use crate::IonCliCommand;
+use crate::commands::IonCliCommand;
 
 use crate::commands::beta::from::json::FromJsonCommand;
