+++ src/bin/ion/commands/dump.rs
@@ -1,4 +1,4 @@
-use crate::IonCliCommand;
+use crate::commands::IonCliCommand;
 use anyhow::{Context, Result};
 use clap::{value_parser, Arg, ArgAction, ArgMatches, Command};
 use ion_rs::*;
